llm_dnd_guessing_game/requirements.txt
python-dotenv
streamlit
openai
